RELEASE: 0.5.3
git-svn-id: https://plugins.svn.wordpress.org/job-manager/trunk@200291 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ jobman-conf.php
@@ -1227,19 +1227,19 @@ function jobman_list_applications() {
 	// Add category filter
 	if( array_key_exists( 'jobman-categories', $_REQUEST ) && is_array( $_REQUEST['jobman-categories'] ) ) {
 		$filtered = true;
-		$args['jobman_category__in'] = array();
+		$args['jcat__in'] = array();
 		foreach( $_REQUEST['jobman-categories'] as $cat ) {
-			$args['jobman_category__in'][] = $cat;
+			$args['jcat__in'][] = $cat;
 		}
 	}
 	
 	// Add star rating filter
-	if( array_key_exists( 'jobman-rating', $_REQUEST ) ) {
+	if( array_key_exists( 'jobman-rating', $_REQUEST ) && is_int( $_REQUEST['jobman-rating'] ) ) {
 	    $args['meta_key'] = 'rating';
 	    $args['meta_value'] = $_REQUEST['jobman-rating'];
 	    $args['meta_compare'] = '>=';
 	}
-	
+
 	$applications = get_posts( $args );
 	
 	$app_displayed = false;
@@ -1578,7 +1578,7 @@ function jobman_application_delete() {
 
 		// Delete any files uploaded
 		foreach( $file_fields as $fid ) {
-			if( array_key_exists( "data$fid", $appdata ) ) {
+			if( array_key_exists( "data$fid", $appdata )  && '' != $appdata["data$fid"] ) {
 				$filename = JOBMAN_UPLOAD_DIR . '/uploads/' . $appdata["data$fid"];
 				if( file_exists( $filename ) )
 					unlink( $filename );
